Moving MS to .net standard
Made a separate matter slice lib and exe
