+++ MatterSliceLib/OptimizedMesh.cs
