+++ MatterSliceLib/SimpleMesh.cs
