+++ MatterSliceLib/utils/PolygonHelper.cs
