+++ MatterSliceLib/utils/PolygonsHelper.cs
