+++ MatterSliceLib/utils/StringHelper.cs
