+++ Program.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace MatterSlice
+{
+	class Program
+	{
+		private static int Main(string[] args)
+		{
+			// this sets the global culture for the app and all new threads
+			CultureInfo.DefaultThreadCurrentCulture = CultureInfo.InvariantCulture;
+			CultureInfo.DefaultThreadCurrentUICulture = CultureInfo.InvariantCulture;
+
+			// and make sure the app is set correctly
+			Thread.CurrentThread.CurrentCulture = CultureInfo.InvariantCulture;
+			Thread.CurrentThread.CurrentUICulture = CultureInfo.InvariantCulture;
+
+			return MatterHackers.MatterSlice.MatterSlice.ProcessArgs(args);
+		}
+	}
+}
