+++ Tests/MatterSlice.Tests/MatterSlice.Tests.csproj
@@ -61,9 +61,9 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\..\MatterSlice.csproj">
-      <Project>{c46ca728-dd2f-4dd1-971a-aaa89d9dff95}</Project>
-      <Name>MatterSlice</Name>
+    <ProjectReference Include="..\..\MatterSliceLib\MatterSliceLib.csproj">
+      <Project>{7f077116-2923-4a77-87cc-ec3be7eb8bc3}</Project>
+      <Name>MatterSliceLib</Name>
     </ProjectReference>
     <ProjectReference Include="..\..\MSClipperLib\MSClipperLib.csproj">
       <Project>{2c564be1-352d-4ddb-8226-f0981f983c60}</Project>
